Inherit test methods from TaxBrainTableResults class
